+++ BlackVision/LibBlackVision/Source/Engine/Models/Plugins/HelperIndexedGeometryConverter.cpp
@@ -300,7 +300,7 @@ IndexedGeometry     IndexedGeometryConverter::MakeIndexGeomFromStrips     ( Floa
     }
 
     // Remove degenerated triangles.
-    for( int i = 0; i < indices.size(); i += 3 )
+    for( int i = 0; i < ( Int32 )indices.size(); i += 3 )
     {
         if( indices[ i ] == indices[ i + 1 ]     ||
             indices[ i + 1 ] == indices[ i + 2 ] ||
